+++ .gitignore
@@ -117,3 +117,4 @@ dist-ssr
 *.sw?
 
 **/front_config.json
+!.github/demoData/front_config.json
