+++ src/front/vite-project/src/types/index.ts
@@ -1,0 +1,2 @@
+export * from  "./record";
+export * from  "./frontConfig";
