+++ src/front/vite-project/src/types/record.ts
@@ -1,0 +1,9 @@
+// バックエンドから返ってくるレスポンスの型
+export interface Record {
+  work_id: string;
+  sub_title: string;
+  number: string;
+  episode_id: string;
+  line: number;
+  body: string;
+}
